requirements.txt
aim==3.16.2
aim-mlflow==0.2.0
aim-ui==3.16.2
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
base58==2.0.1
binaryornot==0.4.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==39.0.2
cycler==0.11.0
databricks-cli==0.17.6
docker==6.0.1
entrypoints==0.4
fastapi==0.95.0
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.8.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mlflow==2.2.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
psutil==5.9.4
py3nvml==0.2.7
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.2
RestrictedPython==6.0
scikit-learn==1.2.2
scipy==1.10.1
segment-analytics-python==2.2.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.1.0
torch==2.0.0
torchmetrics==0.11.4
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
Werkzeug==2.2.3
xmltodict==0.13.0
yarl==1.8.2
zipp==3.15.0
